% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Define a latent random effect in Rusty-INLA}
\usage{
f(covariate, model)
}
\arguments{
\item{covariate}{The variable to map to the latent effect.}

\item{model}{A string defining the model ("iid", "rw1", "ar1").}
}
\value{
A list containing the covariate name and model type.
}
\description{
Define a latent random effect in Rusty-INLA
}
