% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{predict.rusty_inla}
\alias{predict.rusty_inla}
\title{Bayesian Prediction bypass for rusty-INLA}
\usage{
\method{predict}{rusty_inla}(object, newdata, ...)
}
\description{
Automatically appends new data, generates NA targets, forces the NA-trick
through the Rust backend, and extracts the posterior marginal quantiles.
}
