% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rusty_inla}
\alias{rusty_inla}
\title{Native R Formula Interface for Rusty-INLA}
\usage{
rusty_inla(formula, data, family)
}
\arguments{
\item{formula}{A robust R formula \code{y ~ 1 + f(cov, model="iid")}.}

\item{data}{A data.frame containing the variables.}

\item{family}{The likelihood family.}
}
\description{
Native R Formula Interface for Rusty-INLA
}
