% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_inla_run}
\alias{rust_inla_run}
\title{Execute rust-inla backend from R.}
\usage{
rust_inla_run(
  data,
  model_type,
  likelihood_type,
  fixed_matrix_arg,
  n_fixed_arg,
  n_latent_arg,
  x_idx_arg
)
}
\arguments{
\item{data}{A numeric array of length N (the response y).}

\item{model_type}{String identifier ("iid", "rw1", "ar1").}

\item{likelihood_type}{String identifier ("gaussian", "poisson", "gamma").}

\item{fixed_matrix_arg}{Optional flat vector of K columns combined}

\item{n_fixed_arg}{Number of columns K}
}
\description{
Execute rust-inla backend from R.
}
