% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_inla_run}
\alias{rust_inla_run}
\title{Execute rust-inla backend from R.}
\usage{
rust_inla_run(
  data,
  model_types_arg,
  likelihood_type,
  fixed_matrix_arg,
  n_fixed_arg,
  n_latent_arg,
  a_i_arg,
  a_j_arg,
  a_x_arg,
  extr_constr_arg,
  n_constr_arg
)
}
\arguments{
\item{data}{A numeric array of length N (the response y).}

\item{likelihood_type}{String identifier ("gaussian", "poisson", "gamma").}

\item{model_type}{String identifier ("iid", "rw1", "ar1").}

\item{intercept}{Boolean to estimate global beta0.}
}
\description{
Execute rust-inla backend from R.
}
