% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_inla_run}
\alias{rust_inla_run}
\title{Execute rust-inla backend from R.}
\usage{
rust_inla_run(spec)
}
\arguments{
\item{spec}{A named backend specification list built by \code{build_backend_spec()}.}
}
\description{
Execute rust-inla backend from R.
}
